Slight cleanup and enum type/name safety

+++ 3ds/source/gui/screen/InjectorScreen.cpp
@@ -491,7 +491,7 @@ void InjectorScreen::update(touchPosition* touch)
             TitleLoader::save->mysteryGift(*wondercard, tmpSlot);
             if (adaptLanguage)
             {
-                TitleLoader::save->language((u8)lang);
+                TitleLoader::save->language(lang);
             }
             Gui::screenBack();
             return;
